fix logging of events

+++ src/main/java/com/recommender/recommenderAlgorithm/models/Log.java
@@ -9,11 +9,11 @@ public class Log {
     @GeneratedValue(strategy = GenerationType.AUTO, generator = "log_seq")
     @SequenceGenerator(name = "log_seq", sequenceName = "log_seq", allocationSize = 1)
     private Long id;
-    private Integer userId;
+    private Long userId;
     private Integer serialId;
     private String event;
     public Log(){}
-    public Log(Integer userId, Integer serialId, String event) {
+    public Log(Long userId, Integer serialId, String event) {
         this.userId = userId;
         this.serialId = serialId;
         this.event = event;
@@ -27,11 +27,11 @@ public class Log {
         this.id = id;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
